Moved Route53 code examples into route53 directory
