+++ aws/sdk/examples/route53/Cargo.toml
@@ -1,13 +1,17 @@
 [package]
-name = "route53-helloworld"
+name = "route53-code-examples"
 version = "0.1.0"
-authors = ["Alistair McLean <mclean@amazon.com>"]
+authors = ["Alistair McLean <mclean@amazon.com>", "Doug Schwartz <dougsch@amazon.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 route53 = { package = "aws-sdk-route53", path = "../../build/aws-sdk/route53" }
+aws-types = { path = "../../build/aws-sdk/aws-types" }
+
 tokio = { version = "1", features = ["full"] }
+structopt = { version = "0.3", default-features = false }
+tracing-subscriber = { version = "0.2.16", features = ["fmt"] }
 # used only to enable basic logging:
 env_logger = "0.8.2"
